Move the OggVorbis tests to test_ogg.cpp. No need to add test_vorbis.cpp.
