+++ tests/test_ogg.cpp
@@ -16,6 +16,7 @@ using namespace TagLib;
 class TestOGG : public CppUnit::TestFixture
 {
   CPPUNIT_TEST_SUITE(TestOGG);
+  CPPUNIT_TEST(testAudioProperties);
   CPPUNIT_TEST(testSimple);
   CPPUNIT_TEST(testSplitPackets);
   CPPUNIT_TEST(testDictInterface1);
@@ -23,6 +24,17 @@ class TestOGG : public CppUnit::TestFixture
   CPPUNIT_TEST_SUITE_END();
 
 public:
+
+  void testAudioProperties()
+  {
+    Ogg::Vorbis::File f(TEST_FILE_PATH_C("empty.ogg"));
+    CPPUNIT_ASSERT(f.audioProperties());
+    CPPUNIT_ASSERT_EQUAL(3, f.audioProperties()->length());
+    CPPUNIT_ASSERT_EQUAL(3685, f.audioProperties()->lengthInMilliseconds());
+    CPPUNIT_ASSERT_EQUAL(9, f.audioProperties()->bitrate());
+    CPPUNIT_ASSERT_EQUAL(2, f.audioProperties()->channels());
+    CPPUNIT_ASSERT_EQUAL(44100, f.audioProperties()->sampleRate());
+  }
 
   void testSimple()
   {
